Ensure people/projects is not `false` before counting
The UCO functions will return `false` if the requested object
type is the same as the object being displayed. This appeared in
the server logs once because `count( false )` is `1` in PHP 5.6
through PHP 7.1 - https://3v4l.org/Vur8v

I'm not exactly sure how often this would come up, but it doesn't
hurt to be defensive and check for a truthy value before attempting
to count it.